+++ README.md
@@ -23,7 +23,9 @@ please make sure to update to the latest version before using in production.
 
 2021-05-04: v2.1.4 release
 - Added tool to automatically generate spreadsheets with tokens for the test procedure
+- Clarify ambiguity about the re-enabling of PAYG
 - Ensured the example implementations used the recommended values
+- Added full test procedure for the device simulator
 
 2020-10-23: v2.1.1 release
 - Added tool to generate CSV files with device data; the data can then be used for factory setup and software setup
