+++ tests/test_spreadsheet_generator.py
@@ -1,93 +1,91 @@
 import os, sys, inspect
 parentdir = os.path.dirname(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))))
 sys.path.insert(0, parentdir)
-from validation_test.helpers import ADD_TIME, SET_TIME, DISABLE_VALUE, generate_from_device_data, test_accepted_validator, test_how_many_days_validator, test_name
-from simulators.device_simulator import DeviceSimulator
-import codecs
+from tests.helpers import ADD_TIME, SET_TIME, DISABLE_VALUE, generate_from_device_data, test_accepted, test_how_many_days, test_name
 
 
-def run_core_token_tests(device_data, device_simulator):
+def run_core_token_tests(device_data):
 
     test = 'We enter an invalid token'
     token_g1 = '123 456 789'
-    test_accepted_validator(device_simulator, 'G1', token_g1, False, description=test)
+    test_accepted('G1', token_g1, False, description=test)
 
     test = 'We enter a valid token for setting one day'
     device_data, token_g2 = generate_from_device_data(device_data, token_type=SET_TIME, value_days=1)
-    test_how_many_days_validator(device_simulator, 'G2', token_g2, 1, description=test)
+    test_how_many_days('G2', token_g2, 1, description=test)
 
     test = 'We enter a valid token for adding one day'
     device_data, token_g3 = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=1)
-    test_how_many_days_validator(device_simulator, 'G3', token_g3, 2, description=test)
+    test_how_many_days('G3', token_g3, 2, description=test)
 
     test = 'We enter a valid token for setting 30 days and ensures it sets and does not add to the existing'
     device_data, token_g4 = generate_from_device_data(device_data, token_type=SET_TIME, value_days=30)
-    test_how_many_days_validator(device_simulator, 'G4', token_g4, 30, description=test)
+    test_how_many_days('G4', token_g4, 30, description=test)
 
     test = 'We enter a valid token for setting 0 days and ensures the device is inactive with the outputs disabled immediately'
     device_data, token_g5 = generate_from_device_data(device_data, token_type=SET_TIME, value_days=0)
-    test_how_many_days_validator(device_simulator, 'G5', token_g5, 0, description=test)
+    test_how_many_days('G5', token_g5, 0, description=test)
 
     test = 'We enter 3 consecutive Add Time tokens with the maximum amount of days and ensure that they cumulate properly'
     for i in range(1,3+1):
         device_data, token_g6 = generate_from_device_data(device_data, token_type=ADD_TIME, value_raw=995*i)
-        test_how_many_days_validator(device_simulator, test_name('G6', i), token_g6, value_raw=995*i, device_data=device_data, description=test)
+        test_how_many_days(test_name('G6', i), token_g6, value_raw=995*i, device_data=device_data, description=test)
         test = ''
 
     test = 'We enter 21 consecutive Set Time tokens for 1, 2, 3, … 21 days each with a count 30 higher than the other. The validation of the token should not take more than 5 seconds'
     for i in range(1,21+1):
         device_data, token_g7 = generate_from_device_data(device_data, token_type=SET_TIME, value_days=i, token_count=device_data['token_count']+29)
-        test_how_many_days_validator(device_simulator, test_name('G7', i), token_g7, value_days=i, device_data=device_data, description=test)
+        test_how_many_days(test_name('G7', i), token_g7, value_days=i, device_data=device_data, description=test)
         test = ''
 
     test = 'We enter a PAYG Disable token into the device'
     device_data, token_g8 = generate_from_device_data(device_data, token_type=SET_TIME, value_raw=DISABLE_VALUE)
-    test_how_many_days_validator(device_simulator, 'G8', token_g8, None, description=test)
+    test_how_many_days('G8', token_g8, None, description=test)
 
     test = 'We enter a Set Time token for 0 day, it should relock the device'
     device_data, token_g9 = generate_from_device_data(device_data, token_type=SET_TIME, value_days=0)
-    test_how_many_days_validator(device_simulator, 'G9', token_g9, 0, description=test)
+    test_how_many_days('G9', token_g9, 0, description=test)
 
     test = 'We enter a PAYG Disable token to relock the device, then enter a Add Time token with 0 day, it should NOT relock the device (Optional)'
     device_data, token_g10a = generate_from_device_data(device_data, token_type=SET_TIME, value_raw=DISABLE_VALUE)
-    test_how_many_days_validator(device_simulator, 'G10A', token_g10a, None, description=test)
+    test_how_many_days('G10A', token_g10a, None, description=test)
     device_data, token_g10b = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=0)
-    test_how_many_days_validator(device_simulator, 'G10B', token_g10b, None)
+    test_how_many_days('G10B', token_g10b, None)
 
     test = 'We deactivate the device with a Set Time of 0 days. We then wait 48 hours before entering a Add Time of 1 day and ensuring that the days late are not considered in the activation time'
     device_data, token_g11a = generate_from_device_data(device_data, token_type=SET_TIME, value_days=0)
-    test_how_many_days_validator(device_simulator, 'G11A', token_g11a, 0, description=test)
+    test_how_many_days('G11A', token_g11a, 0, description=test)
     device_data, token_g11b = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=1)
-    test_how_many_days_validator(device_simulator, 'G11B', token_g11b, 1)
+    test_how_many_days('G11B', token_g11b, 1)
     
     return device_data
 
 
-def run_unordered_entry_tests(device_data, device_simulator):
+def run_unordered_entry_tests(device_data):
     
     test = 'We generate 3 Add Time tokens, then enter the 3rd, then first, then second and ensure the days are added properly'
     device_data, token_u1a = generate_from_device_data(device_data, token_type=SET_TIME, value_days=60)
-    test_how_many_days_validator(device_simulator, 'U1A', token_u1a, 60, description=test)
+    test_how_many_days('U1A', token_u1a, 60, description=test)
     device_data, token_u1b = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=10)
     device_data, token_u1c = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=5)
     device_data, token_u1d = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=1)
-    test_how_many_days_validator(device_simulator, 'U1B', token_u1d, 61)
-    test_how_many_days_validator(device_simulator, 'U1C', token_u1b, 71)
-    test_how_many_days_validator(device_simulator, 'U1D', token_u1c, 76)
+    test_how_many_days('U1B', token_u1d, 61)
+    test_how_many_days('U1C', token_u1b, 71)
+    test_how_many_days('U1D', token_u1c, 76)
 
     test = 'We generate an Add Time, a Set Time and another Add Time token. We enter the set time and ensure that the older Add Time does not work but the newer does'
     device_data, token_u2a = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=5)
     device_data, token_u2b = generate_from_device_data(device_data, token_type=SET_TIME, value_days=10)
     device_data, token_u2c = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=3)
-    test_how_many_days_validator(device_simulator, 'U2A', token_u2b, 10, description=test)
-    test_accepted_validator(device_simulator, 'U2B', token_u2a, False)
-    test_how_many_days_validator(device_simulator, 'U2C', token_u2c, 13)
+    test_how_many_days('U2A', token_u2b, 10, description=test)
+    test_accepted('U2B', token_u2a, False)
+    test_how_many_days('U2C', token_u2c, 13)
 
     test = 'We generate an Add Time token and a Disable PAYG token, we enter the Disable PAYG token and then the Add Time token should be refused'
     device_data, token_u3a = generate_from_device_data(device_data, token_type=ADD_TIME, value_days=1)
-    device_data, token_u3b = generate_from_device_data(device_data, token_type=ADD_TIME, value_raw=DISABLE_VALUE)
-    test_how_many_days_validator(device_simulator, 'U3A', token_u3b, None, description=test)
-    test_accepted_validator(device_simulator, 'U3B', token_u3a, False)
+    device_data, token_u3b = generate_from_device_data(device_data, token_type=SET_TIME, value_raw=DISABLE_VALUE)
+    test_how_many_days('U3A', token_u3b, None, description=test)
+    test_accepted('U3B', token_u3a, False)
 
     return device_data
 
@@ -96,15 +94,13 @@ def run_unordered_entry_tests(device_data, device_simulator):
 
 if __name__ == '__main__':
     device_data = {
-        'serial_number': 'HQ0000AAAAA',
+        'serial_number': 'changeme',
         'starting_code': 123456789,
-        'key': '4c2bdd66b78a8d0b1b78b82314f3f531',
+        'key': 'example',
         'restricted_digit_set': False,
         'time_divider': 1,
-        'token_count': 1
+        'token_count': 2
     }
-    device_simulator = DeviceSimulator(device_data['starting_code'], codecs.decode(device_data['key'], 'hex'),
-                                       restricted_digit_set=device_data['restricted_digit_set'],
-                                       waiting_period_enabled=False)
-    device_data = run_core_token_tests(device_data, device_simulator)
-    device_data = run_unordered_entry_tests(device_data, device_simulator)
+    print('Test Name,Token Used,Expected Result,Test Details,Result')
+    device_data = run_core_token_tests(device_data)
+    device_data = run_unordered_entry_tests(device_data)
